Add files via upload
all bug is fixed.
visual summary and diffculty is now displayed in the stduy score tracker 
calculation is now included in testscore.py application as it saves more performance and more easiler to monitor the change of score. 
Probably the final edition for the basic application.
All down with just me and Chatgpt 
Thanks!